refactor(Add comment): changed adding comments, removed props made an editor request, errors were removed from the console
