+++ src/components/AddComments/interfaces.ts
@@ -1,7 +1,3 @@
-export interface IAddComponentProps {
-  onSubmitForm: (value: string) => void;
-}
-
 export interface ISubmitValue {
   value: string;
 }
